Improve CORS checking for casting to Chromecast
1. When checking whether the video is playable send headers that
match Chromecast's headers.
2. Check the contents of returned CORS headers, if any, to see if
it is compatible with the Chromecast origin.

Note that this CL doesn't change when the video is checked for
playability, so videos with bad CORS will still show the cast
button, but will not be sent to the Chromecast.

BUG=539406

Review URL: https://codereview.chromium.org/1378693003

Cr-Original-Commit-Position: refs/heads/master@{#352836}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3fda2269814ef8e0af4210adb4d30706f78cc8c