Add suppressions for crbug.com/545273.
Not sure at this point if this is a Dr. Memory bug or a V8 bug. There
was an update to Dr. Memory when this issue started to show up and no
change to V8 though there was a roll 2 builds previous and this issue is
slightly flaky.

BUG=545273
TBR=bruening@chromium.org

Review URL: https://codereview.chromium.org/1414883002

Cr-Original-Commit-Position: refs/heads/master@{#354950}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d23e5227812017c98c1d4fec6228cfec84a1c91c