Extracts more information from serial devices on Mac
This is a modified reland of crrev.com/1306733008.

Originally, I completely replaced the old enumeration method with the
new one. After someone spotted a regression for the Windows code, I
reverted the enumeration code that yielded more detailed device info
and am now recommitting it as a hybrid of the old enumeration method
and new one. We merge the lists of new devices and old devices, giving
preference to the new ones (with more detailed info). I also added an
UMA metric to track the difference in the number of devices that each
method yields. When that metric shows that the two methods are always
returning the same number of devices, we can eliminate the old method
entirely.

BUG=522217

Review URL: https://codereview.chromium.org/1399023002

Cr-Original-Commit-Position: refs/heads/master@{#355161}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 65e7fba107d7af4b65e8b4af4eaec0b04423bb29