[Telemetry] Allows concatenating multiple Disabled & Enabled decorators.
For this to work, we also abandon the "@Disabled" &
"@Disabled()" syntax, which used to mean unconditionally
disable. The places that use such syntax are switched
to @Disabled('all').
The reason for this is to simplify the handling logic of
combining multiple decorators & make the unconditional
@Disabled(..) clearer.

BUG=552477

Review URL: https://codereview.chromium.org/1428133005

Cr-Original-Commit-Position: refs/heads/master@{#358457}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f83d36c71ece47e23fac753b42437148090636c1