Preparation work for third_party/openh264 build-from-src CL
This CL reverts the parts of https://codereview.chromium.org/1394983003/ that added dummy unittest binaries for openh264.
This was decided to be a bad move (overhead of binaries, only wanting the tests if flag is set). The unittests will be placed and implemented in content_unittests instead in a follow-up CL.

This CL also pulls in openh264/src source code. The building of this and test code will be added in a follow-up CL.

BUG=500605, 468365

Review URL: https://codereview.chromium.org/1438553002

Cr-Original-Commit-Position: refs/heads/master@{#359548}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc28e3ffd89faa7cf1a38ca63084685186534108