Add an experiment to switch to an existing tab.
This is experiment controls the behaviour of what happens when a user opens a
suggestion from the NTP. Specifically if there are previously opened tabs for
the suggestion.
We are starting this with 2 simple cases where we match the existing tabs and
the suggestion by URL or by Hostname.

This is a base change so we can understand which direction to proceed from here.

BUG=565953

Review URL: https://codereview.chromium.org/1501753002

Cr-Original-Commit-Position: refs/heads/master@{#364319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04ca405a4c0ca62806f6493e91d60b7fc0924747