Add Pre-read Experiment: Preread Only Cold DLLs.
A Chrome process's main DLL is pre-read on each launch on Windows. This
CL adds the EnabledColdOnly group to the PreRead field trial. For users
in this group, DLLs will be pre-read only if they are cold. To detect
that a DLL is cold, we read some bytes from its first page and check if
that generates a hard fault.

This should improve the following metrics:
- PLT.PT_RequestToFinish, by about 5% at the 50th percentile.
- Startup.FirstWebContents.NonEmptyPaint2.WarmStartup, by about 8% at
  the 50th percentile.
- Startup.BrowserMessageLoopStartHardFaultCount, half the number of hard
  faults at the 50th percentile.
- Less cold startups, because Superfetch does a better job?
(based on results obtained with the "Disabled" group)

BUG=547794

Review URL: https://codereview.chromium.org/1499223002

Cr-Original-Commit-Position: refs/heads/master@{#364350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4072fb31c998f310dc67a6928e4470f67b0eef18