Add more stats for tracking strict secure cookies deletion behavior
This adds UMA stats for tracking the removal of old cookies because of a
set cookie operation. Records all attempts to delete an equivalent
cookie in a set cookie operation as well as when it successfully finds
and removes one. Also measures when a cookie is skipped for deletion
because of strict secure cookies rules.

BUG=567867

Review URL: https://codereview.chromium.org/1531473005

Cr-Original-Commit-Position: refs/heads/master@{#365670}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31e32630bd0fba4e46561e7ce8165710f54f4670