Return no responses in IterResponses if timeline model is None.
In some tests we don't get any traffic, which causes an excpetion in
IterResponses when self._timeline_model is None when it's not expected
to be.  In particular this is causing a test to fail when an extra via
header is expected.

BUG=576308

Review URL: https://codereview.chromium.org/1578823002

Cr-Original-Commit-Position: refs/heads/master@{#368677}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80bbad867c3627ed6e44bf4dcebe74b6b3a182c6