roll clang 255169:256316
New:
* /W4 in clang-cl enables -Wextra warnings
* fix for link.exe /incremental build flakes
* /Brepro suppresses timestamps for deterministic builds
  (doesn't work with link.exe /incremental)

Note that this brings in a libc++ regression on Mac [1]: some of the
symbols that are hidden symbols in the system libc++ in OS X 10.8
are marked as "inline" in the headers shipped with this compiler --
this means that binaries produced by this compiler might not run
on OS X 10.8 if these calls don't get inlined (or, if the 10.8 SDK is
used, they won't link). That's ok for us though because production
builds on OS X statically link a newer libc++ with these symbols
exported, the iOS bots run newer OS X versions than 10.8, and
so do the Mac bots that do component builds (which use the
system libc++) or asan builds. And soon we'll roll in r257193 which
fixes this again.

1: http://lists.llvm.org/pipermail/cfe-commits/Week-of-Mon-20160104/146428.html

BUG=571691
TBR=hans

Review URL: https://codereview.chromium.org/1539373005

Cr-Original-Commit-Position: refs/heads/master@{#368796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 483df9304bb678dd50b435a41c063a3aed96d10f