Changes frame length to mean bytes of samples in frame
I misunderstood the meaning of the frame header's "length" field: I
thought that it meant the number of samples in the frame. It actually
means the number of bytes of samples in the frame. This CL fixes that
mistake.

BUG=542837

Review URL: https://codereview.chromium.org/1579753002

Cr-Original-Commit-Position: refs/heads/master@{#368863}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf22c76e62e2798fc6ebbfd37bfe101f7c0a6e05