GCM: Remove ChromeGcmReceiver on Android
The ChromeGcmReceiver was added to avoid crashes due to SecurityException
thrown in the GcmReceiver when trying to start the GcmListenerService due
to a framework bug. This has now been fixed in the GCM client library
which catches the exception.

BUG=580367

Review URL: https://codereview.chromium.org/1629433002

Cr-Original-Commit-Position: refs/heads/master@{#371266}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c72d5c71a4d195268fbc119b20df8874b80c9a8