Implement the Windows desktop search redirection feature.
With this CL, Windows desktop searches are redirected to the default
search engine when the DesktopSearchRedirection feature is enabled.
The first time that a desktop search is redirected, an infobar is
shown to inform the user of the redirection.

BUG=539963

Review URL: https://codereview.chromium.org/1598553003

Cr-Original-Commit-Position: refs/heads/master@{#373049}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2e7c477ce837b177ed32eaa7a2fd1e7911b67a87