tools/android/loading: Speed up trace events processing.
The previous way to index trace events allowed for O(log n) query of
in-flight events at a given time, and O(n) constrction time, but at the
cost of replicating events up to p times (p being the maximum number of
concurrent events). It turns out that p is fairly high.

This replaces this with a tree, which is worse, but in practice faster.
On a typical loading trace, creating the PNG graph goes from 3m30s to ~20s.

Review URL: https://codereview.chromium.org/1694223002

Cr-Original-Commit-Position: refs/heads/master@{#375575}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1570bad784b3dd9fa1daf0bf125635a2b77f08cb