tools/android/loading Enable async call stacks
This CL enables the "Debugger" domains and sets the async
call stack depth to 4 for the request track.
The asynchronous stacks are merged with the basic stacks, which
provides more information in the initiators and dramatically reduces
the number of orphan requests.

Review URL: https://codereview.chromium.org/1692983002

Cr-Original-Commit-Position: refs/heads/master@{#375870}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b382e4f33c93cf1e95c6e351f365f72ef8b27f54