Upstream filtering of libraries not used on iOS from generated credits.
Chrome on iOS only uses a subset of the libraries used by Chromium
so add an extra argument to tools/licenses.py corresponding to the
OS targetted by the current build.

This is used on iOS to filter the libraries that are not used by
Chrome on iOS from the generated credits page.

Upstream this temporary solution until the more general solution
from the original bug as this is blocking the effort to unfork
Chrome on iOS and the team responsible for the unforking does not
have the manpower to implement the full solution.

BUG=178215

Review URL: https://codereview.chromium.org/1701743002

Cr-Original-Commit-Position: refs/heads/master@{#375921}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d13d15b0c45e1d456c45f2874b92d9de22dd81cb