[win/cros/lin] Add back the spellcheck submenu.
This patch adds back the "Spellcheck" submenu to the context menu on
Windows, ChromeOS, and Linux. This partially reverts commit
http://crrev.com/5414139041953f2b558cfb0fdf8a4927a19a3144, but also adds
support for multilingual spellcheck in the submenu via the "All
languages" radio button. Selecting this radio button will spellcheck in
all of the languages that are listed in the submenu. The language list
can be modified in chrome://settings/languages.

TBR=brettw@chromium.org
BUG=544234

Review URL: https://codereview.chromium.org/1647723002

Cr-Original-Commit-Position: refs/heads/master@{#375955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e59900a51f10891ee99368eda83ddec185f2b007