Add various UMA histograms for measuring GCM crypto performance
This adds (and tests) UMA in the GCMKeyStore, an on-disk database, and
the decryption routines part of the GCMEncryptionProvider and the
GCMMessageCryptographer.

Adding the metrics comes with an extra free refactoring adjusting the
message-received flow to make it cleaner to measure them.

BUG=486040

Review URL: https://codereview.chromium.org/1707513002

Cr-Original-Commit-Position: refs/heads/master@{#376501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 266a2aa40f2fb11662afe0af694b4be0cfb3efe4