Remove incorrect PLT.PT.NavigationStartToFirstLayout histogram
This is a revert of https://codereview.chromium.org/1230513002

The logic for this histogram incorrectly assumes that a RenderViewObserver's
lifetime is tied to a single page load. This is incorrect: a RVO is tied to the
lifetime of a RenderView, which can live for many page loads. Thus, the logic to
compute and emit this histogram is incorrect. We've since added new page load
histograms under PageLoad.*, and there is a
PageLoad.Timing2.NavigationToFirstLayout which is correctly emitted once per
page load. Thus, we are deprecating PLT.PT.NavigationStartToFirstLayout in favor
of PageLoad.Timing2.NavigationToFirstLayout.

BUG=593018

Review URL: https://codereview.chromium.org/1774873002

Cr-Original-Commit-Position: refs/heads/master@{#380418}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 267cd99070e74fc51e38d278f5a557275064c21d