Instrument MojoShellConnectionImpl::BindToMessagePipe
This adds a histogram to measure the duration of calls to
MojoShellConnectionImpl::BindToMessagePipe in child processes.

The new histogram replaces a useless one which was recently added
to measure uninteresting calls to MojoShellConnection::Create()
in renderers.

BindToMessagePipe cost should be significantly reduced in a
follow-up CL, but it will still need to do some non-trivial
amount of work. This metric will help us detect if it remains
problematic or regresses.

BUG=591742
R=ben@chromium.org

Review URL: https://codereview.chromium.org/1782133002

Cr-Original-Commit-Position: refs/heads/master@{#380555}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7b3190c0ecd5b88688aa5e55d3085f0fbd40e11