Add histograms to track down the cause of ERR_SSL_PROTOCOL_ERROR.
These histograms will be removed when the bug is resolved (or really as soon
as we get data back, so hopefully Monday...). Notably, they record BoringSSL
error reasons which are *not* stable.

The intent of these histograms is to:

- Determine what BoringSSL error is triggering these.

- Based on magnitudes, confirm this is coming from Read/Write and not Connect.

- See if there is a correlation with some cipher suite by comparing the
  distribution with the normal cipher suite distribution.

BUG=593963

Review URL: https://codereview.chromium.org/1786593003

Cr-Original-Commit-Position: refs/heads/master@{#380827}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e50b9f54970b1527162fba5611b8d1b41a9d9ba