Make DrMemory update process reliable
The DrMemory update process only runs when it thinks that it needs to
but its heuristic is not accurate. A better thing to do is to always
update.

This change also adds setlocal so that environment variables don't
leak outside. This may also help with making this script more robust
in the future, once it has been in place for long enough for all
machines affected by the environment variable leakage to have rebooted.

BUG=595867
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1807013004 .

Cr-Original-Commit-Position: refs/heads/master@{#381870}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38e785b06f3942adad154fddd1b83d8a0f299a5e